Fixed pong getting stuck
I've been testing a few things but in the old code it got stuck occasionally on the top of the screen. The only kind of thing I could think of that could be was the pong was going too fast for the collision to detect for the edges, so I changed it so the expectation takes in the pongs speed into consideration for the top. for the bottom I just included the expectation of the height and speed to be a factor when deciding how low the pong can get before bouncing back up.