fixed pong getting stuck above near the top.
after fiddling around more some with the code, my theory behind the ball getting stuck was the comparison between a float and an integer, thus I set the float to become an integer for ease of use, though making the speed of the ball vary a lot less. helps for a more enjoyable experience.